Added a .gitignore that hides all of the generated output files from git status and the like.
Added jDOM back as an Ivy dependency.  I'm not sure why this was removed.

Added a new command line interface nade RetrieveStoriesForIteration.  Some
liberal refactoring between RetrieveStories and RetrieveStoriesByID was
necessary to make this nice and clean.

Added tasks and task lists to the model and XML representation.  I started
adding the plumbing for task retrieval as well.  The XSLT stuff needs to be
written yet.