Added explicitly handling for errors and warnings in responses.  Both are logged and any error results in aborting result processing.  This could probably be strengthened into an exception.
Added preliminary handling for xsi:type attributes.  I posted to agile commons
about adding xsi:type usage in responses.  Basically the "fetch=true" option
that returns "full" child objects instead of links to them is pretty much
worthless when using JAXB.  It will not *cast* them to the appropriate type
unless a proper xsi:type attribute is included.  The trick that I use is to
check the actual type with an `instanceof` in processQueryResult.  If I have
the type that I am looking for, then I can simply use it.  Otherwise, I have
to go and make another query to retrieve the actual object.

Rewrote the query creation code to use an improved algorithm and removed all
traces of the `includeFullObjects` option since it is broken for the time
being.

Made some hefty modifications in the tests package.  Instead of having inline
XML documents in the test code, I moved them into `configs/test-data`.  All of
the test XML content is completely new.  I stole a bunch of example content off
of http://agilesoftwaredevelopment.com/scrum/simple-product-backlog and
encoded it into Rally XML response documents.  There is still more to do here
if I want to push the test coverage up, but this is a nice start.