Made a bunch of build related changes and some XML schema changes.  The schema changes aren't that visible to the outside world.  I added a new abstract type named "top-level-object" that contains nothing more than a set of links.  I could pull more commonality into this type in the future.  The main reason for the schema change is to make it possible to abstractly add hypermedia links to our model objects (e.g., those in standup.xml).
I went through and made use of the links collection in ServerConnection.  Now
nearly every instance of a model class includes a bunch of links that track
where it came from and related objects.  This is the first step to fully
supporting the up-and-coming RESTful web service.

The Ant script was modified to separate the test code and data out into a
separate jar.  This change required a bunch of carefully created exclude and
include filesets.  This is also on the road to using the guts of this as a
web service via JAX-RS.