Moved generated sources from the "real-clean" target into "dist-clean" for no particular reason except that I kept forgetting to regen the sources.
Added methods to HttpClientFactory that are responsible for creating the
HTTP request objects.  The Rally REST API recommends that a number of custom
headers be set on requests.  The easiest way to do this was to provide an
entry point that controls that creation of HTTP request objects.  This
required a bunch of changes including a new class for Rally specific
connections.