Started adding the web frontend using JavaServer Faces.  I did a lot more code reoganization in the process.
The properties files are now stored in subdirectories in the ``configs`` tree
that parallel the subdirectories in the ``src`` tree.  This makes deployment a
lot cleaner since I just apply the configurations in a strict order -
main -> test -> web.  The newer files simply overwrite the older ones.  For
example, when building the test JAR file, we copy all of the properties from
configs/main, then properties from configs/test.  Clashing names will be
overwritten - this makes specific configuration of log4j quite easy to manage.

As a side note, I managed to get log4j to write to both a custom log file
stored in *.../domains/.../logs/application.log* as well as having it write
to the global *server.log* file.  I'm not sure if I like the idea of cluttering
up server.log with application specific stuff, but it will work for now.
This can be easily removed by tweaking the ``configs/web/log4j.properties``
file appropriately.

I also created all of the necessary WEB-INF stuff in ``src/web/WEB-INF``.  This
will serve as a model for customized META-INF directories if they are ever
needed.  It might be just as appropriate to stash this stuff in the ``configs``
tree... need to think about that a little more.

The last thing that I did was write two simple servlets and a Faces backing
bean.  I'm pretty sure that the servlets are unnecessary and will be removed
shortly.  The backing bean exposes the rally.ServerConnection class with some
extra methods needed for validation.  This will form the core of the web
application ... or at least that is the current plan.