Moved all of the non-source code into the configs tree.  This includes all of the web pages, XML schemas, XSL transforms, and XML test data.  Now the configs tree essentially mimics the src tree in directory structure.  I am already having a change of heart about this, but I'm going to commit this structure for safe keeping.
I was forced to add "anyAttribute" support for every element that can be
buried in the description element.  I also had to add support for <font>
elements being just about anywhere in the description as well.  It looks
like Rally is pretty accepting on which HTML elements they support.  The
direction that I took is to pass them through into the intermediate
representation and then strip them out whilst generating the FO.  This is
going to have to change since I really don't want to clutter up the
intermediate XML representation... but that will have to wait for a little
while.

The last real change was the addition of the "--debug-prefix" command line
option to the CLI tools.  This will cause the application to dump out some
of the intermediate XML files.  I needed this while I was trying to figure
out why the XML transformation of <description> was failing miserably.