Modified RetrieveStories to prompt for information.
If the user & password are not supplied, then the program prompts
for them.  I will use System.console to prompt for a visually secure
password if possible but will fall back to using System.in if need
be.

Also fixed some ugliness in the XSL that was causing owners to be
omitted from stories.